% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nginar.R
\name{error.nginar}
\alias{error.nginar}
\title{Function error.nginar}
\usage{
error.nginar(n, alpha, mu)
}
\arguments{
\item{n}{the length of outputs series. A strictly positive integer.}

\item{alpha}{a vector of INAR coefficients.}

\item{mu}{a vector.}
}
\value{
Resultados
}
\description{
Fit in inar model to a univariate time series by yule-walker method
}
\examples{

\dontrum{
function (blah)
}
}

