% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innovations.R
\name{innovation}
\alias{innovation}
\alias{print.innovation}
\title{Family of Innovations Process for Fitting an INAR(p) Model}
\usage{
innovation(dist)

\method{print}{innovation}(x, ...)
}
\arguments{
\item{dist}{Character specification of the innovation process, see
details.}

\item{x}{An \code{"innovation"} object.}

\item{...}{Further arguments for other specific methods.}
}
\value{
The function \code{innovation()} returns an \code{"innovation"} object
    that brings a set of information about the innovation process . More
    specifically, returns a list with the following
    elements:
 \itemize{
   \item{d:}{ Probability mass function \code{function(x, par)}.}
   \item{r:}{ Random generator function \code{function(n, par)}.}
   \item{par:}{ Describes the parametric space of the distribution.}
   \item{npar:}{ Number of parameters of the innovation process specified.}
   \item{constraint:}{ Character; describes the restriction between
       parameters if any.}
   \item{disp:}{ Describes the types of dispersion that the distribution
       can model.}
   \item{name:}{ Name of the distribution.}
 }

 The arguments of the returned functions are:
 \describe{
   \item{\code{x}}{ Vector of discrete non-negative quantiles.}
   \item{\code{n}}{ Number of observations to return.}
   \item{\code{par}}{ Parameter vector of the innovation process.}
 }
}
\description{
Provide the current available distributions that can be
    used as an innovation process in a fit of the INAR(p) model.
}
\details{
There are some discrete distributions available for the
    innovation process specification. The following table display their
    names and their abbreviations to be passed to \code{innovation()}.

  \tabular{llll}{
 \bold{Distribution}  \tab \bold{Abbreviation} \tab \tab \bold{Parameters}\cr
 Bernoulli \tab \code{"BE"} \tab \tab \code{0 < theta < 1} \cr
 BerPoi    \tab \code{"BP"} \tab \tab \code{theta > 0; 0 < phi < 1} \cr
 BerG    \tab \code{"BG"} \tab \tab \code{theta, phi > 0} \cr
 Geometric \tab \code{"GE"} \tab \tab \code{theta > 0} \cr
 Mean-Parameterized COM-Poisson \tab  \code{"CP"} \tab \tab \code{theta, phi > 0} \cr
 Negative Binomial \tab \code{"NB"} \tab \tab \code{theta, phi > 0} \cr
 Poisson  \tab \code{"PO"}      \tab \tab  \code{theta > 0}  \cr
 }
}
\examples{
\dontrun{
 ### Specification of the Poisson innovation to 'inv' object
 inv <- innovation("PO")

 ### Methods
 inv

 ### Generating observations
 x <- inv$r(500, 5)

 ### Barplot and probability mass function
 xaxis <- barplot(prop.table(table(x)), main = inv$name,
                  xlab = "x", ylab = "Proportion")
 points(xaxis, inv$d(sort(unique(x)), 5),
        type = "b", pch = 16, col = 2)
}

}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
