% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninarsimm.R
\name{nginar.sim}
\alias{nginar.sim}
\title{Function nginar.sim}
\usage{
nginar.sim(n, alpha, mu, n.start = 150)
}
\arguments{
\item{n}{A strictly positive integer.}

\item{alpha}{a vector of INAR coefficients.}

\item{mu}{a vector.}

\item{n.start}{the length of 'burn-m' period. If na, the default, a reasonable valve is computed.}
}
\value{
Resultados
}
\description{
definir
}
\examples{

}

