% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar1.R
\name{poinar}
\alias{poinar}
\title{Function poinar}
\usage{
poinar(x, order.max, series = NULL)
}
\arguments{
\item{x}{a numeric vector or time series.}

\item{order.max}{a one dimensional integer vestor giving the order of the model to fit. This value corresponds the INAR order.}

\item{series}{name for the series.}
}
\value{
Resultados
}
\description{
Fit in inar model to a univariate time series by yule-walker method
}
\examples{

\dontrum{
function (blah)
}
}

