% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inarsim.R
\name{poinar.sim}
\alias{poinar.sim}
\title{Function poinar.sim}
\usage{
poinar.sim(n, order.max, alpha, lambda, n.start = NA)
}
\arguments{
\item{n}{the length of outputs series. A strictly positive integer.}

\item{order.max}{the integer component p is the INAR order.}

\item{alpha}{a vector of INAR coefficients.}

\item{lambda}{the mean of the poisson distribution.}

\item{n.start}{the length of 'burn-m' period. If na, the default, a reasonable valve is computed.}
}
\value{
For \code{inar} and its methods a list of class \code{"inar"} with the following elements:
}
\description{
definir
}
\examples{

\dontrum{
function (blah)
}
}

