% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar_simmulation.R
\name{simulation}
\alias{simulation}
\alias{inar.sim}
\title{Simulate From an INAR(p) Model}
\usage{
inar.sim(n, alpha, par, inv = "PO", n.start = NA)
}
\arguments{
\item{n}{A strictly positive integer given the length of the output series.}

\item{alpha}{A vector of INAR coefficients.}

\item{par}{A vector with the innovation process parameters
specified in the following order: mean and dispersion (or precision).}

\item{inv}{Character specification of the innovation process, see
details.}

\item{n.start}{The length of 'burn-in' period. If \code{NA},
the default, a reasonable value (500) is computed.}
}
\value{
A time-series object of class \code{"ts"}.
}
\description{
Simulate from an INAR(p) model.
}
\details{
There are some discrete distributions available for the
    innovation process specification. The following table display their
    names and their abbreviations to be passed to \code{\link[tsinteger]{innovation}()}.

 \tabular{lll}{
 \bold{Distribution}  \tab \bold{Abbreviation} \tab \bold{Number of parameters}\cr
 BerPoi    \tab \code{"BP"} \tab 2 \cr
 BerG    \tab \code{"BG"} \tab 2 \cr
 Geometric \tab \code{"GE"} \tab 1 \cr
 Poisson  \tab \code{"PO"}      \tab  1  \cr
 }
}
\examples{
# A Poisson INAR(3) simulation
y <- inar.sim(n = 1000, alpha = c(0.2, 0.3, 0.3), par = 5)
layout(matrix(c(1,2,1,2,1,3,1,3), ncol = 4))
plot(y)
acf(y, main = "")
pacf(y, main = "")
layout(1)

}
\references{
Du, J.G. and Li,Y. (1991).
  The integer-valued autorregressive (INAR(p)) model.
  \emph{Journal of time series analysis}. \bold{12}, 129--142.
}
