% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innovations.R
\name{innovation}
\alias{innovation}
\alias{print.innovation}
\title{Family of Innovations Process for Fitting an INAR(p) Model}
\usage{
innovation(dist)

\method{print}{innovation}(x, ...)
}
\arguments{
\item{dist}{Character specification of the innovation process, see
details.}

\item{x}{An \code{"innovation"} object.}

\item{...}{Further arguments for other specific methods.}
}
\value{
The function \code{innovation()} returns a list whose components
    are set of functions referring to the innovation process. More
    specifically, returns an \code{"innovation"} object with the following
    elements:
 \itemize{
   \item{d:}{ Probability mass function \code{function(x, mu, phi)}.}
   \item{r:}{ Random generator function \code{function(n, mu, phi)}.}
   \item{npar:}{ Number of parameters of the innovation process specified.}
   \item{constraint:}{ Character; describes the restriction between
       parameters if any.}
   \item{disp:}{ Describes the types of dispersion that the distribution
       can model.}
   \item{name:}{ Name of the distribution.}
 }

 The arguments of the returned functions are:
 \describe{
   \item{\code{x}}{ Vector of discrete non-negative quantiles.}
   \item{\code{n}}{ Number of observations to return.}
   \item{\code{mu}}{ Mean parameter of the innovation process.}
   \item{\code{phi}}{ Dispersion parameter of the distribution, if it
    does not exist, it receives \code{NULL}.}
 }
}
\description{
Provide the current available distributions that can be
    used as an innovation process in a fit of the INAR(p) model.
}
\details{
There are some discrete distributions available for the
    innovation process specification. The following table display their
    names and their abbreviations to be passed to \code{innovation()}.

  \tabular{lll}{
 \bold{Distribution}  \tab \bold{Abbreviation} \tab \bold{Number of parameters}\cr
 Bernoulli \tab \code{"BE"} \tab 1 \cr
 BerPoi    \tab \code{"BP"} \tab 2 \cr
 BerG    \tab \code{"BG"} \tab 2 \cr
 Mean-Parameterized COM-Poisson \tab \code{"CP"} \tab 2 \cr
 Geometric \tab \code{"GE"} \tab 1 \cr
 Poisson  \tab \code{"PO"}      \tab  1  \cr
 }
}
\examples{
\dontrun{
 ### Specification of the Poisson innovation to 'inv' object
 inv <- innovation("PO")

 ### Methods
 inv
 plot(inv)

 ### Generating observations
 x <- inv$r(500, 5)

 ### Barplot and probability mass function
 xaxis <- barplot(prop.table(table(x)), main = inv$name,
                  xlab = "x", ylab = "Proportion")
 points(xaxis, inv$d(sort(unique(x)), 5),
        type = "b", pch = 16, col = 2)
}

}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
